auth & work with models

+++ trello/task_app/views.py
@@ -1,20 +1,19 @@
+from django import forms
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 import json
 
+# CRUD
+# List objects | GET
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse, JsonResponse
 from django.shortcuts import get_object_or_404, redirect, render
-from django.views.decorators.csrf import csrf_exempt
-
-from .models import Task, Comment
-
-# CRUD
-
-# List objects | GET
-
-
-from django.contrib.auth.decorators import login_required
 from django.utils.decorators import method_decorator
-
+from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_POST
+
+from .models import Comment, Task
+
 
 # Детали таска + комментарии
 @login_required
@@ -53,22 +52,27 @@ def task_delete(request, task_id):
     return redirect("task_list_or_create")
 
 # Post Create Object | POST
+
+
 def task_list_or_create(request):
     print(request.POST)
     if request.method == 'POST':
         title = request.POST.get('title')
         desc = request.POST.get("description")
+        user = request.user
         if title:
             # при создании всегда попадает в "todo" в конец
-            max_pos = Task.objects.filter(status="todo").count()
-            Task.objects.create(title=title, desc=desc, status="todo", position=max_pos)
+            max_pos = Task.objects.filter(owner=user, status="todo").count()
+            Task.objects.create(title=title, desc=desc,
+                                status="todo", position=max_pos, owner=user)
 
         return redirect('task_list_or_create')
-    
+
     #     Task.objects.create(title=title, desc=desc)
     #     return redirect('task_list_or_create')
-
-    tasks = Task.objects.all().order_by("position")
+    # if request.user.is_authenticated:
+    tasks = Task.objects.filter(owner=request.user).order_by("position")
+    # tasks = Task.objects.all().order_by("position")
     return render(request, 'task_list.html', {'tasks': tasks})
     # tasks = Task.objects.all().order_by('-id')
     # return render(request, 'task_list.html', {'tasks': tasks})
@@ -76,14 +80,15 @@ def task_list_or_create(request):
 
 def task_detail(request, task_id):
     task = Task.objects.get(id=task_id)
-    return render(request, 'task_detail.html', {'task': task})
+    return render(request, 'task_detail.html', {'task': task, 'comments': task.comments.all()})
 
 
 def task_delete(request, task_id):
     task = get_object_or_404(Task, id=task_id)
+    if task.owner != request.user:
+        return HttpResponse("You are not authorized to delete this task.", status=403)
     task.delete()
     return redirect('task_list_or_create')
-
 
 
 @csrf_exempt
@@ -94,18 +99,16 @@ def update_task_status(request):
         new_status = data.get('new_status')
         order = data.get("order", [])
 
-        # Print task_id and new_status for debugging
-        print(f"Task ID: {task_id}, New Status: {new_status}")
-
-        # Update the task's status
         try:
+            # обновляем статус задачи
             task = Task.objects.get(id=task_id)
             task.status = new_status
             task.save()
 
-             # обновляем порядок в колонке
+            # обновляем позиции задач в колонке
             for item in order:
-                Task.objects.filter(id=item["id"]).update(position=item["position"])
+                Task.objects.filter(id=item["id"]).update(
+                    position=item["position"])
 
             return JsonResponse({
                 "message": "Task status & order updated!",
@@ -122,10 +125,6 @@ def update_task_status(request):
             return JsonResponse({"error": "Task not found!"}, status=404)
 
     return JsonResponse({"error": "Invalid request!"}, status=400)
-    #         return JsonResponse({"message": "Task status updated successfully!"}, status=200)
-    #     except Task.DoesNotExist:
-    #         return JsonResponse({"error": "Task not found!"}, status=404)
-    # return JsonResponse({"error": "Invalid request!"}, status=400)
 
 
 @csrf_exempt
@@ -140,11 +139,71 @@ def add_comment(request, task_id):
     text = data.get("text", "").strip()
     if not text:
         return JsonResponse({"error": "Empty comment"}, status=400)
-
+    print(task_id, data)
     comment = Comment.objects.create(task=task, author=request.user, text=text)
+    print(comment)
     return JsonResponse({
         "id": comment.id,
         "author": comment.author.username,
         "text": comment.text,
         "created_at": comment.created_at.strftime("%Y-%m-%d %H:%M"),
     })
+
+from django import forms
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.models import User
+
+#     username = forms.CharField(max_length=150)
+#     email = forms.EmailField(required=False)
+#     password = forms.CharField(widget=forms.PasswordInput)
+#     password2 = forms.CharField(widget=forms.PasswordInput, label="Confirm Password")
+
+
+def register(request):
+    if request.method == "POST":
+        username = request.POST.get("username")
+        email = request.POST.get("email")
+        password = request.POST.get("password")
+        password2 = request.POST.get("password2")
+
+        if len(username) < 3:
+            return render(request, "register.html", {"error": "Username must be at least 3 characters long."})
+
+        if len(password) < 6:
+            return render(request, "register.html", {"error": "Password must be at least 6 characters long."})
+
+        if password != password2:
+            return render(request, "register.html", {"error": "Passwords do not match."})
+
+        if username and password:
+            try:
+                User.objects.get(username=username)
+                return render(request, "register.html", {"error": "Username already exists."})
+            except User.DoesNotExist:
+                User.objects.create_user(
+                    username=username, password=password, email=email)
+                return redirect("login")
+        else:
+            return render(request, "register.html", {"error": "Please provide both username and password."})
+
+    return render(request, "register.html")
+
+
+def user_login(request):
+    if request.method == "POST":
+        username = request.POST.get("username")
+        password = request.POST.get("password")
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect("task_list_or_create")
+        else:
+            return render(request, "login.html", {"error": "Invalid credentials."})
+
+    return render(request, "login.html")
+
+
+def user_logout(request):
+    logout(request)
+    return redirect("login")
